UpdateMenu, User, GameSession,UserScoreMAnager
GameSession - To put user name and score to write into the file. // TODO. to complete update score into a file.

Menu - update method and create loop for to provide menu. But it is still not working properly like user can put menu method. // TODO, to make it completely as a limitation for user's input.

UserScoreManager - creating a method for updating user name and score.